refactor: isolate MySQL read-only policy
